pyproject.toml
[project]
name = "document-extractor"
version = "0.1.0"
description = "Multi-modal document data extractor using LangGraph for entity and relation extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-groq>=0.1.0",
    "langchain-community>=0.3.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "pillow>=10.0.0",
    "docling>=1.0.0",
    "networkx>=3.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytesseract>=0.3.10",
    "matplotlib>=3.10.8",
    "neo4j>=5.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
